Ticket: #5107 - MPR series when initially generated shows incorrect direction marker on the text overlay '''Schema/Configuration Changes: None'''  * Added overloads to ImageOrientationPatient to allow getting the secondary direction while accounting for a small amount of tolerance - when the corresponding directional cosine is within the specified tolerance, the function will return None.  * DirectionalMarkerAnnotationItems now use a 1-degree tolerance when getting the secondary direction for display.  * VolumeSliceSopDataSource now sets the ImageOrientationPatient components using SetFloat32 rather than SetFloat64 to prevent an automatic, precision-losing upconversion (the data computation was performed using 32-bit floats).

+++ ImageViewer/Volume/Mpr/VolumeSliceSopDataSource.cs
@@ -81,12 +81,12 @@ namespace ClearCanvas.ImageViewer.Volume.Mpr
 
 			// assign Image Orientation (Patient)
 			Matrix resliceAxesPatientOrientation = _volume.Volume.RotateToPatientOrientation(_resliceMatrix);
-			_instanceDataSet[DicomTags.ImageOrientationPatient].SetFloat64(0, resliceAxesPatientOrientation[0, 0]);
-			_instanceDataSet[DicomTags.ImageOrientationPatient].SetFloat64(1, resliceAxesPatientOrientation[0, 1]);
-			_instanceDataSet[DicomTags.ImageOrientationPatient].SetFloat64(2, resliceAxesPatientOrientation[0, 2]);
-			_instanceDataSet[DicomTags.ImageOrientationPatient].SetFloat64(3, resliceAxesPatientOrientation[1, 0]);
-			_instanceDataSet[DicomTags.ImageOrientationPatient].SetFloat64(4, resliceAxesPatientOrientation[1, 1]);
-			_instanceDataSet[DicomTags.ImageOrientationPatient].SetFloat64(5, resliceAxesPatientOrientation[1, 2]);
+			_instanceDataSet[DicomTags.ImageOrientationPatient].SetFloat32(0, resliceAxesPatientOrientation[0, 0]);
+			_instanceDataSet[DicomTags.ImageOrientationPatient].SetFloat32(1, resliceAxesPatientOrientation[0, 1]);
+			_instanceDataSet[DicomTags.ImageOrientationPatient].SetFloat32(2, resliceAxesPatientOrientation[0, 2]);
+			_instanceDataSet[DicomTags.ImageOrientationPatient].SetFloat32(3, resliceAxesPatientOrientation[1, 0]);
+			_instanceDataSet[DicomTags.ImageOrientationPatient].SetFloat32(4, resliceAxesPatientOrientation[1, 1]);
+			_instanceDataSet[DicomTags.ImageOrientationPatient].SetFloat32(5, resliceAxesPatientOrientation[1, 2]);
 
 			// assign Image Position (Patient)
 			Vector3D topLeftOfSlicePatient = GetTopLeftOfSlicePatient(frameSize, throughPoints[0], volume, slicerParams);
